BWT RLE compression
Implements BWT RLE compression and decompression
