+++ Test Files/RLE_Test_BWT.txt
@@ -1,0 +1,1 @@
+pzAAABBBBCCCCCDDDDDDEEEEEEEFFFFFFF|Fhijklmno^xy
