+++ Test Files/RLE_Test_BWT_RLEcompr.txt
@@ -1,0 +1,1 @@
+11p1z3A4B5C6D7E7F1|1F1h1i1j1k1l1m1n1o1^1x1y1
