+++ Test Files/RLE_Test_BWT_RLEcompr_RLEdecomp.txt
@@ -1,0 +1,1 @@
+pzAAABBBBCCCCCDDDDDDEEEEEEEFFFFFFF|Fhijklmno^xy
