+++ Test Files/RLE_Test_BWT_RLEcompr_RLEdecomp_BWTinvert.txt
